[spaces.interface] update sparsity pattern creation
This commit uses recent additions in the sparsity pattern in dune-stuff,
namely making use of the `insert(ii, jj)` method in order to not duplicate
the manual search everywhere. In addition the pattern is now sorted upon
return to be on the safe side. If the performance penalty is too large we
will have to introduce an additional optional switch a la
`compute_pattern(const bool sort = true)`.