[spaces.constraints] introduced CRTP interface
Unfortunately there is no way around this: ConstraintsInterface now is a
CRTP interface. Otherwise template matching of the local_constrains()
method of a space would break, since the wrapper of the constraints used
in the system assembler would erase the derived type. New constraints can
easily be added with a minimum amount of boiler-plating, though, using
Constraints::internal::Default as a base, see Constraints::Dirichlet for
example.