[products] complete rewrite, WARNING: this will creak your code!
This commit introduces a complete rewrite of the products. If you are just
interested in the consequences for using products: see the next commit to
get an idea of the changes you need to make in your code.

As before products/base.hh provides LocalizableBase, AssemblableBase and
now GenericBase. These classes fully implement their respective
Interfaces given a helper class LocalOperatorProvider. This class has to
be provided for each type of products, i.e. once for an L2 product (see
products/l2-internal.hh), once for an elliptic product
(see products/elliptic-internal.hh), ... This class has to provide a
protected member local_operator_ which is used in the base classes and some
types that are used in the base traits in products/base-internal.hh. For
each type of products one has to add a small class combining this
local operator provider with the appropriate base product, see for instance
L2Localizable or L2 in products/l2.hh. The local operator provider classes
are really necessary since they are the ones creating the arguments to the
local operator (the local operator could have been created in the base
class but the necessary data functions are unknown there). The final
product classes combining the local operator provider with the base product
is not strictly necessary. But since those are around twelve lines of code
and very simple I would like to keep them to have a good place for
documentation (especially for documenting the respective ctors, the
arguments of which might not always be straightforward). I did not try to
maintain any backwards compatibility or header forwards since some things
will break because of this anyway: given the implementation of the base
products and their forward to the local operator provider all arguments
for the base have to be given first, i.e L2Weighted(grid_view, factor).
In addition:
* renamed H1SemiGeneric -> H1Semi to match the other products
* moved elliptic product from playground
* changed the order of ctor arguments (s.a.): always the ones for the base
  product first, i.e. in Products::Elliptic:
  diffusion, grid_view -> grid_view, diffusion
* changed the order of template arguments to match the ctor arguments
  where possible, i.e. in Products::Elliptic:
  DiffusionType, GridViewType -> GridViewType, DiffusionType
* over_integrate has to be given to the ctor now and not to apply2()