[spaces...fem] hold shared_ptr of grid part (again)
This fixes a segfault, since after a copy of the space the backend_ would
still hold a reference to the grid_part_ of other which is not guaranteed
to still exist. Thus we would need to make a copy of backend_ (not possible
with dune-fem) or create a new backend_ (which would mathematically
hopefully be a copy, but we do not know about dune-fems internals so we can
not be sure) or again hold a shared_ptr of the backend_ and the grid_part_
and the grid_view_, which I opted for.