[local] refactor (WARNING: THIS MIGHT BREAK YOUR CODE!)
The goal is to make dune-gdt less template-heavy, so the local operator
and functional interfaces are now virtual (as opposed to the old CRTP-based
ones). This greatly simplifies the whole library code, but comes with the
additional price that more information has to be provided on construction
of these objects (so the user has to pay the price).