[test.linearelliptic] refactor eoc grid provider usage
The EocGridProvider as the base of each test case is now instantiated with
a grid config, not a grid. This allows for more general eoc grid providers.
The EocGridProvider, on the other hand, now provides level_provider(...) for
each refinement (as opposed to the layer directly), which is a GridProvider
to be passed on to the discretizer.