[P|operators] refactor bindings structure
The OperatorInterface requires the ConstLincombOperator, the
LincombOperator and the MatrixBasedOperator and correspondingly binds all
these types now. They can be split over multiple compile units, since no
overloaded function in a single module is desired. From the matrix operator
we only want the factory method in a single module, so we split the type
bindings from the factory bindings.